fix(if): resolve if.html error
fixes https://github.com/aurelia/templating-resources/issues/312
